refactor: asset handling and add renderer extensions
Move asset compilation, locating and optimization logic into dedicated classes (Asset\Compiler, Asset\Locator, Asset\Optimizer) and update Asset to use them; deprecate legacy static helpers in favor of Locator/Slugifier. Add BuildContextInterface and make Builder implement it to reduce coupling. Introduce new Twig extensions (Renderer\Extension\Collection and Renderer\Extension\Content) for collection/content filters and functions. Replace Page::SLUGIFY_PATTERN and slugify implementation to use new Util\Slugifier. Add image optimization and minify delegation, and ensure remote asset downloads respect a configurable timeout (default added to config). Update tests and phpunit to include Unit tests suite. Minor related adjustments across Generator, Renderer, Step and Util classes to integrate the refactor.
